requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
cointables==0.1.0
contourpy==1.0.7
cycler==0.11.0
dateparser==1.1.8
DateTime==5.1
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
mplfinance==0.12.9b7
multidict==6.0.4
numpy==1.24.2
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pycryptodome==3.17
pyparsing==3.0.9
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
urllib3==1.26.15
websockets==11.0.1
yarl==1.8.2
zipp==3.15.0
zope.interface==6.0
